install.sh: implement workaround for resume from hibernation with swapfile
systemd initrd currently attempts to load the root filesystem (/sysroot)
in parallel with attempting to resume from the saved state. The kernel
prevents this as there is a high likelihood of data corruption when
attempting to mount the same resource at once. This commit requires the
root filesystem to begin mounting after local-fs-pre.target and
therefore after the system checks for a saved state.