install.sh: add comment about HibernateDelaySec regression
HibernateDelaySec was updated to no longer hibernate after a specified
time period under sleep, but instead check every unit of the specified
amount of time to see if the battery is low (<5%) and subsequently enter
hibernation. This was disliked and reverted in favor of a new option
while also returning the original functionality of HibernateDelaySec to
the next release of Systemd, version 253.