tests/fido*.nix: updated testScripts "wait_for_console_text" prompt value.
Fixes an issue with the FIDO test scripts where the test would fail when
waiting for the tty to prompt for the backup password. For some reason,
the previous password prompt "Please enter recovery key for disk fido"
was no longer recognized although it still appears on the console tty.
The prompt has now been updated to match the text that appears before
that prompt: "Starting password query on".