script.nix: explicitly specify package versions in script
As the script depends on the user having the latests version of packages
available on the nixpkgs-unstable repository, making sure that those
version of packages are used instead of the system defaults in
essential. Previously, this was done through specifying dependencies in
'shell.nix' and requiring the user to run 'nix-shell'. Now the
dependencies are explicitly specified in the 'script.nix' file.