package.nix: remove build inputs
Remove build inputs from Nix package.

In scenarios where the installation script is used on a machine with
limited disk and memory space (i.e. a virtual machine), there would not
be enough space present for all of the build inputs to be installed and
therefore the installation script could not be compiled.

However, the build inputs are already included in the installation ISO
and therefore do not **need** to be installed in order for the
installation script to work. The only problems that may arise are

	1. The ISO used does not have one/some of the programs present
	2. The version of the program that is installed in the ISO is
	   older than what is in the flake.lock nixpkgs version **and**
	   missing some options that this script uses.

In order to resolve the first issue, a check can be added to the start
of the program in order to check that the programs are all present. The
second issue is more difficult to resolve.